edit docs, refactor race tab
