+++ package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "esbuild": "^0.23.0",
+    "tanstack-table-export-to-csv": "^1.0.5",
     "uuid": "^11.0.5"
   },
   "devDependencies": {
