+++ packages/client/src/components/config/lapTableConfig.ts
@@ -1,0 +1,75 @@
+import { IFormattedLapData } from "@shared/helios-types";
+import { createColumnHelper } from "@tanstack/react-table";
+
+const columnHelper = createColumnHelper<IFormattedLapData>();
+
+export const columns = [
+  columnHelper.accessor("data.timeStamp", {
+    cell: (info) => info.getValue(),
+    header: "Time Stamp",
+    sortingFn: (rowA, rowB, columnId) => {
+      const parseDate = (dateString: string) => {
+        const parts = dateString.split(" ");
+
+        const dateParts = parts[0]?.split("/");
+        const timeParts = parts[1]?.split(":");
+
+        if (!dateParts || !timeParts) {
+          throw new Error("Invalid Time");
+        }
+        const hour = Number(timeParts[0]);
+        const minute = Number(timeParts[1]);
+
+        const day = Number(dateParts[1]);
+        const month = Number(dateParts[0]);
+        const year = Number(dateParts[2]);
+        return new Date(year, month - 1, day, hour, minute).getTime();
+      };
+      const dateA = parseDate(rowA.getValue(columnId));
+      const dateB = parseDate(rowB.getValue(columnId));
+
+      return dateB - dateA;
+    },
+  }),
+  columnHelper.accessor("data.ampHours", {
+    cell: (info) => info.getValue(),
+    header: "Amp Hours",
+  }),
+  columnHelper.accessor("data.averagePackCurrent", {
+    cell: (info) => info.getValue(),
+    header: "Average Pack Current",
+  }),
+  columnHelper.accessor("data.averageSpeed", {
+    cell: (info) => info.getValue(),
+    header: "Average Speed",
+  }),
+  columnHelper.accessor("data.batterySecondsRemaining", {
+    cell: (info) => info.getValue(),
+    header: "Battery Seconds Remaining",
+  }),
+  columnHelper.accessor("data.distance", {
+    cell: (info) => info.getValue(),
+    header: "Distance",
+  }),
+  columnHelper.accessor("data.energyConsumed", {
+    cell: (info) => info.getValue(),
+    header: "Energy Consumed",
+  }),
+  columnHelper.accessor("data.lapTime", {
+    cell: (info) => info.getValue(),
+    header: "Lap Time",
+  }),
+  columnHelper.accessor("data.netPowerOut", {
+    cell: (info) => info.getValue(),
+    header: "Net Power Out",
+  }),
+
+  columnHelper.accessor("data.totalPowerIn", {
+    cell: (info) => info.getValue(),
+    header: "Total Power In",
+  }),
+  columnHelper.accessor("data.totalPowerOut", {
+    cell: (info) => info.getValue(),
+    header: "Total Power Out",
+  }),
+];
